requirements.txt
Pillow==5.1.0
urllib3==1.22
setuptools==28.8.0